requirements.txt
fastapi
uvicorn
duckdb
asyncpg
pandas
pyyaml
google-cloud-storage
fsspec
gcsfs
